Syntibs/requirements.txt
openpyxl
pyyaml
lark
nltk
pdf2image
data
tempdir
python-docx

